+++ application/extractor/extractor_test.go
@@ -25,74 +25,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// member.go
-func TestPublicKeyResponse(t *testing.T) {
-	testValue := "test_public_key"
-
-	data, err := core.Serialize([]interface{}{testValue, nil})
-	require.NoError(t, err)
-
-	result, err := PublicKeyResponse(data)
-
-	require.NoError(t, err)
-	require.Equal(t, testValue, result)
-}
-
-func TestPublicKeyResponse_ErrorResponse(t *testing.T) {
-	testValue := "test_public_key"
-	contractErr := &foundation.Error{S: "Custom test error"}
-
-	data, err := core.Serialize([]interface{}{testValue, contractErr})
-	require.NoError(t, err)
-
-	result, err := PublicKeyResponse(data)
-
-	require.Contains(t, err.Error(), "Has error in response")
-	require.Contains(t, err.Error(), "Custom test error")
-	require.Equal(t, "", result)
-}
-
-func TestPublicKeyResponse_UnmarshalError(t *testing.T) {
-	testValue := "some_no_valid_data"
-
-	data, err := core.Serialize(testValue)
-	require.NoError(t, err)
-
-	result, err := PublicKeyResponse(data)
-
-	require.Contains(t, err.Error(), "Can't unmarshal")
-	require.Equal(t, "", result)
-}
-
-func TestCallResponse(t *testing.T) {
-	testValue := map[interface{}]interface{}{
-		"string_value": "test_string",
-		"int_value":    uint64(1),
-	}
-
-	data, err := core.Serialize([]interface{}{testValue, nil})
-	require.NoError(t, err)
-
-	result, contractErr, err := CallResponse(data)
-
-	require.NoError(t, err)
-	require.Nil(t, contractErr)
-	require.Equal(t, testValue, result)
-}
-
-func TestCallResponse_UnmarshalError(t *testing.T) {
-	testValue := "some_no_valid_data"
-
-	data, err := core.Serialize(testValue)
-	require.NoError(t, err)
-
-	result, contractErr, err := CallResponse(data)
-
-	require.Contains(t, err.Error(), "Can't unmarshal response")
-	require.Nil(t, contractErr)
-	require.Nil(t, result)
-}
-
 // node.go
 func TestNodeInfoResponse(t *testing.T) {
 	testPK := "test_public_key"
